Size conversion warning reduction.

+++ eeschema/sch_reference_list.h
@@ -291,8 +291,8 @@ public:
 
     size_t GetCount() const { return m_flatList.size(); }
 
-    SCH_REFERENCE& GetItem( int aIdx ) { return m_flatList[aIdx]; }
-    const SCH_REFERENCE& GetItem( int aIdx ) const { return m_flatList[aIdx]; }
+    SCH_REFERENCE& GetItem( size_t aIdx ) { return m_flatList[aIdx]; }
+    const SCH_REFERENCE& GetItem( size_t aIdx ) const { return m_flatList[aIdx]; }
 
     void AddItem( const SCH_REFERENCE& aItem ) { m_flatList.push_back( aItem ); }
